Move baseline authorizations in domain
Part of epic #12142 Baseline v1

Objective: check authorizations as soon as possible.
Some authorizations are verified when entering in domain (with domain services).
Indeed, domain services are exposed through REST api, but other interfaces may be
possible. So REST interface is not a good candidate for authorizations checks.

Other authorizations are verified in adapters when check cannot be done earlier.
Ex: with findById, we have to query database to known which project is involved.
It is only once you have the instance, that you know the project.

For maintenance reasons, a single component is responsible for all domain authorization
rules: BaselineAuthorizations.

Note: this change is only about authorizations on baseline resource. Authorizations
on comparions are out of scope for now.

Change-Id: I568aa533db5f5eb79dac1564a50a96f6c6a92541