Memberships to private without restricted project are removed when a user becomes restricted
Switching a user to restricted remove him/her to all the projects that are in
the private without restricted mode.
A confirmation is requested by a modal when the user is going to be removed
from at least one project.

The purge is done in a system event as it might take a while.

To test, you will need to enable the "private without restricted" mode by setting
the global parameter feature_flag_project_without_restricted to 1.

This is part of story #13237: Prevent restricted to be added to private projects

Change-Id: I98c3e9a1471411dade5fa8ac65fc2b95c85b524d