Add a content hash on the ressources of the TLP UI framework
Resources of the framework are now built exclusively with Webpack
and each generated asset has a content hash. This make sure the assets
retrieved by the browsers will be up to date and can be cached aggressively.
nginx configuration has been updated to leverage that.

The stylesheets of the FlamingParrot theme are also built with Webpack due
to a direct dependency to the TLP font. The build with the proposed approach
also make sure that all resources linked by the stylesheets exist (the build
crashing otherwise). The FP theme was referencing non existing resources until
this contribution.

The assets built in dev/watch mode now have an inlined source map and are not
minified to ease debug sessions.

No functionnal change is expected.

Part of story #13237: Prevent restricted to be added to private projects

Change-Id: Id62c07ab193a46c8693ecbcaec05e9eab8463233