Refactor baselines loading
Part of epic #12142 Baseline v1

For each baseline displayed in baselines list, we need to load
some extra data:
	- user
	- baseline artifact (and corresponding tracker)

Data fetching is done through baselines presenter and presented
baslines are returned with additional attributes.

This pattern brings many difficulties:
	- the presenter code base is not easy to maintain (not
	trivial complexity)
	- the presenter code base is not easy to test (too many
	thing are done here)
	- returned object is kind of baseline, but not exatcly.
	This implies difficulties for naming such objects in tests.
	- it is very difficult (if not impossible) to share
	fetched data with other components (like futures
	comparisons list).

For all of these reasons, a refactoring is necessary.
Here is the new architecture:
	- all loaded data are stored (and shared) in a global store
	- each fetchable resource has a dedicated action (to
	load the resource) and a dedicated getter (used with
	resource id).
	- additional actions are responsible for loading resources
	and associated ones.
	- when a component need an associated object, it get it
	from the global store with the id

This refactoring is limited to baselines list for change size reason.
So, many trade-off are made, in particular in baseline store action.

Change-Id: I02bf7da22a3081c5a885336866e2a35a37e1146f