Accept both deleteproject and delete-project gerrit plugin
Make sure you've installed delete-project plugin on gerrit 2.14 or 2.16.
Go to tuleap » project » git » repo » settings » gerrit.
If the repository is migrated to gerrit, then it should propose the
option to delete the project from gerrit, like for gerrit 2.13.

This is due to a renaming of the plugin after gerrit 2.13:
- https://gerrit.googlesource.com/plugins/delete-project/+/refs/heads/stable-2.13/BUCK
- https://gerrit.googlesource.com/plugins/delete-project/+/refs/heads/stable-2.14/BUILD

Fix request #13306: gerrit 2.14+ delete-project plugin

Change-Id: I5410b26efed702e9931a802c32c11f0592921408