Introduce a setting allowing to use all project visibility mode
A new setting default_project_visibility can be used with the same
values than the ones available in the XML import (private, private-wo-restr,
unrestricted and public) to set the default project visibility mode.

This setting, when set, takes priority on the existing sys_is_project_public
which will be removed from the default local.inc.

default_project_visibility will be modifiable in the site administration web UI
and through the Tuleap CLI.

Part of story #13237: Prevent restricted to be added to private projects

Change-Id: I3fafd4f284cc2be45352edb06cdacd252a8a42fc