request #13314 Tuleap EventManager should be PSR-14 compliant
The change is a bit useless by itself. However fairly small and it's worth it to:
- makes Tuleap code more "Modern PHP" compatible
- allow to stop events (despite code being dead ATM)
- decouple our event dispatching mechanism from EventManager implementation.

In order to be fully compliant with PSR-14, EventManager now accepts to dispatch
any kind of object (event without a NAME const). It might be used for plugins as
well but will require a bit more boilerplate.

Change-Id: Ib17ccadc80e0b2c03beb997f9debbd65adf903e7