request #13321: Download of large attachment/file can be partial/resumed
This contribution introduce the support of range requests [0]
when the BinaryFileResponseBuilder is used to generate an HTTP
response.

Requests requesting multiple ranges or specifying only how many
bytes from the end of the file to return are not supported. In those
cases no content-range is specified and the full body message is returned
as allowed by the specification. Browsers or frequently encountered tools
does not use them.

Contribution can be tested by observing what's returned when pausing/resuming
a download in a browser or by crafting requests with curl.

[0] https://tools.ietf.org/html/rfc7233

Change-Id: I701c7434cdb39313192e56e983d74a895b493fbd