Optimize http queries in comparison page
Part of epic #12142 Baseline v1

Objective: reduce http queries by regrouping all queries on same
baseline on same artifact depth.

So, all artifacts of first depth level are fetch in a single query
(for a given baseline).
Another query is done for the second level, another for third,
and so on.

Before this patch, there was one query for each artifact which
have linked artifacts.

This was made possible by splitting artifacts fetching from
baseline comparison (with statistics computation).
But this implies that all artifacts are fetch even if they
are not visible (ex: artifacts linked to new or removed
artifacts).
This drawback seems to be acceptable when we see the
code simplification and we don't know yet if this would implies
a performance issue or not.

Change-Id: I0792ff263be9bf3ff33b1e3b316d4bc31ea369fd