PUT create empty FrozenFields post action
This change is part of request #12189

On PUT request with type "frozen_fields" to
/tracker_workflow_transitions/{id}/actions an empty
FrozenFields PostAction is created.

Any other previous FrozenFields PostAction is deleted.

If workflow is not in simple mode, an error is returned.

You need the sys_should_use_read_only_post_actions feature flag to
enable it. Otherwise, the standard behaviour (no frozen_fields type)
is applied.

Change-Id: Ia9a9e267b1c793489e6b4c1f519cf01df9a8d042