Mention in the warning message that the removal of restricted users is not immediate
When switching a project to private without restricted, the removal of the restricted
users is not immediate. The removal is done with system events.
Please note however, that has soon that the project does not accept restricted users,
the restricted users can not access it anymore even if they are no yet completely
removed from it.

This is part of story #13237: Prevent restricted to be added to private projects

Change-Id: Ife864e604bd61a41eb8a1fbe2a692b1fe2870a4f