Select frozen fields in PUT REST route
This is part of story #12189 prevent field modification according to state

A new mandatory key 'field_ids' must be provided in the json content for
frozen fields action in PUT tracker_workflow_transitions/:id/actions

This attribute is an array of integer, that must not be empty.

Only used fields in the tracker can be provided, or a 400 is raised and
nothing is done.

The selected field ids are not yet saved in database.
You still need to defined the feature flag to be able to defined frozen
field action in REST.

Change-Id: I74ec892ef81f211f719cac945e1a59450ea12805