Disallow selecting the workflow field as a frozen one
This is part of story #12189 prevent field modification according to state

If a field id provided in the field_ids attribute is the field used to
define the workflow, the REST route throws an error.

The selected field ids are not yet saved in database.
You still need to defined the feature flag to be able to defined frozen
field action in REST.

Change-Id: I765c9198f7e0ad85858affe2abfc5bc6b8248922