request #13361: Require the nullability symbol to be present when the default value is null
It is more explicit and this discussion [0] from two years ago from
the doctrine/repo also points out it can hide a LSP violation.

Now seems to be a good time to introduce this rule as we start to aggressively type
scalar values so we might encounter the case more often.

To test, you can review the code of the sniff [1], apply the new ruleset and allow
to autofix this rule then launch `make phpcbf`. The diff should be identical with
this contribution.

[0] https://github.com/doctrine/common/pull/805#discussion_r128762923
[1] https://github.com/slevomat/coding-standard/blob/5.0.4/SlevomatCodingStandard/Sniffs/TypeHints/NullableTypeForNullDefaultValueSniff.php

Change-Id: Ia7477be1ee75b40e02ccd4d097b9de3fcca64e4c