Disallow selecting fields with field dependencies for Frozen fields post action
This is part of story #12189 prevent field modification according to state

If a field id provided in the field_ids attribute is a field that has
field dependencies, the REST route throws an error.

The selected field ids are not yet saved in database.
You still need to defined the feature flag to be able to defined frozen
field action in REST.

Change-Id: I05c1f829d406565529b34e68f164fb55b040c4aa