Save Frozen fields actions in database
This is part of story #12189 prevent field modification according to state

Provided field ids are now saved in database if the provided ids are
usable.

You still need to defined the feature flag to be able to defined frozen
field action in REST.

This post action is also duplicated when creating a new transition in
simple mode if the state has a fozen fields action.

Change-Id: Id6afdd285acce3983e734d02bed42809ee7e7636