request #13366: Gitolite user created by Tuleap packaging is not a system user
Only new installations are affected by this changes: there is no good/safe way
to handle this change during an upgrade and it only has consequences for a small
parts of the deployed instances.
Legacy Git plugin has been kept as is, it should not be used for
new installations.

See [0] for the Fedora guidelines for packages requiring the creation of
users and groups.

[0] https://fedoraproject.org/wiki/Packaging:UsersAndGroups

Change-Id: Ia134c4fafe9d4c1e8fa27ef6d7e16c165ce2b13f