First step of simple_workflog XML import
This is part of story #13243 Import/export workflow simple mode via XML

A first base of workflow simple XML import is done.
It allows you to import a workflow in simple mode with basic
informations (the used field and the is used information).

Example:
  <simple_workflow>
    <field_id REF="F42704"/>
    <is_used>1</is_used>
  </simple_workflow>

Neither transitions nor actions can be imported for now.

Workflow and simple workflow cannot be imported at the same time.

Change-Id: I92737367afd1d0e6600617820a40e6d503116f72