Import transitions without actions and conditions in simple mode
This is part of story #13243 Import/export workflow simple mode via XML

Transitions can now be imported for workflows in simple mode.
Example:
<simple_workflow>
    <field_id REF="F42704"/>
    <is_used>1</is_used>
    <states>
      <state>
        <to_id REF="V36863"/>
        <transitions>
            <transition>
                <from_id REF="null"/>
            </transition>
            <transition>
                <from_id REF="V36864"/>
            </transition>
        </transitions>
      </state>
    </states>
  </simple_workflow>

We can defined in XMl a state with the destination value id. Then each
transitions for this state can be defined by adding the from value id.

Change-Id: I79fb8b6e7e1b65a5cb0aa4efc258bc437efd8ccd