request #13376: Artifact modal can not open an artifact with a field without value on which a workflow is defined
In some cases artifacts might not have a value for a field, as a result
the field will not be present when calling GET /artifacts/{id}.
When this field without value is used in a workflow, the modal will not be
able to initialize itself.

This contribution only make sure is a bit more resilient in this situation as
I have not been able to pinpoint all the cases where it can happen. The only
idnetified situation for now is with the import/export XML of an artifact
already in this case.

Change-Id: I890cee1583fa663a503a318a681acf75583fc624