Generate thumbnail for uploaded images
When the upload is finished, we should generate a thumbnail if the file
is an image, just like when we use the web ui to add attachment.

The thumbnail can be reached via (for example):
/plugins/tracker/attachments/preview/123-Lenna.png
(original file being /plugins/tracker/attachments/123-Lenna.png)

The thumbnail is removed when the corresponding file is expired and the
daily cleanup is run.

Part of story #12964: drag'n drop images in tracker text areas

Change-Id: Id67d9d7633a842ce46d10b6f0cde613fbec81132