Begin to export simple workflows in XML
This is part of story #13243 Import/export workflow simple mode via XML

When you export a tracker with a simple workflow, the workflow is now
exported with the simple XML structure.
The export only deals with field_id and is_used for now.

To enable this, you have to set the following feature flag in the
local.inc file: $sys_tracker_export_simple_workflows = 1;

Change-Id: I969eb1343b6ca4f0cc3c0d02b694356312679e38