Refactoring: extract the save of File value
In order to enforce SRP, move logic of field file saveValue() into
dedicated objects, with a better test coverage (before: 0% ☺).

This is a preliminary step to be able to attach tus-uploaded-files to an
artifact. It will be then easier to add new behavior by adding component
in the attachment creator chain of responsibility:

-> try to add a tus-uploaded-file [*]
-> if not, try to add a rest-temporary-file
-> if not, try to add a traditional-upload-file

Part of story #12964: drag'n drop images in tracker text areas

[*] in a next changeset

Change-Id: I529668d4dd518f697d1fec9a91c4cc8d23600815