Export states for XML simple workflows
This is part of story #13243 Import/export workflow simple mode via XML

When you export a tracker with a simple workflow, the workflow is now
exported with the states. There are neither transitions nor
actions/conditions yet.

To enable this, you have to set the following feature flag in the
local.inc file: $sys_tracker_export_simple_workflows = 1;

Change-Id: Icead910f6a266945c46c32352ab86a0e4235ce3d