Export transitions for XML simple workflows
This is part of story #13243 Import/export workflow simple mode via XML

When you export a tracker with a simple workflow, the workflow is now
exported with the states and transitions.
There are neither action nor conditions exported yet.

To enable this, you have to set the following feature flag in the
local.inc file: $sys_tracker_export_simple_workflows = 1;

Change-Id: I9a4154e2bcd9a069409a1f56c0adaf73b20a2e19