getFirstTransitionForCurrentState must return the first non new transition
This is part of story #13243 Import/export workflow simple mode via XML

This is a refactoring needed for the XML export of pre conditions.
Because the condition "comment not empty" does not exist for transitions
that come from "new artifact", we must retrieve the first transition
that does not come from "new artifact" to export in the state all the
pre conditions possible.

With this patch, no functional changes are expected.

Change-Id: Ifc3af82595f49d0503af45f69956dc3654496ece