Update vimeo/psalm to 3.2.12
The baseline has recreated from scratch to take into account
the latest bugfixes of Psalm.

A small fix has been done in forumml_utils.php, the error is
somehow not added to the baseline. It's probably due to the
way the forumml_utils code is included/loaded. Unless we encounter
the issue on a regular matter it is probably not worth to
troubleshoot it.

Next contribution will enforce the static analysis in the CI test
pipeline. It's not done here because we need a bit of extra scripting
to consider only the changed files covered by the Psalm configuration.

Part of request #13292: Integrate vimeo/psalm to the test pipeline

Change-Id: I065c7bb4deef3113059b721ddf6a712df65564c0