Make sure Psalm knows how to access classes of the FlamingParrot theme
Without this contribution if you try to launch Psalm in the LDAP plugin
(`make psalm FILES=plugins/ldap/include/`), you will get issues because
the FlamingParrot_Theme class cannot be found.

Took the opportunity to move the theme to an autoloader managed by Composer
like it is done for the BurningParrot theme.

Part of request #13292: Integrate vimeo/psalm to the test pipeline

Change-Id: I44f25a92678878881a105cb2357db8cccfe6ef6f