tuleap-cfg can run systemctl commands
This replace the service management that was done in shell for install script
by a tuleap-cfg dedicated command. Not only this allows to have code with
tests but it's also easier to set execute a command that does nothing when
systemd is not part of the target system.

This is part of request #13394 tuleap-aio should run on centos7

Change-Id: I318667c48867ee8fad3a660aa5e24614393e2167