Add Psalm verification in the test pipeline on modified files
Only the modified files that are in scope of the Psalm configuration
are analyzed. We might miss some issues in the non modified files
but it allow us to do the analysis quickly without needing to new
resources or increase the pipeline execution time.

Recents errors introduced in the codebase are also fixed.

Part of request #13292: Integrate vimeo/psalm to the test pipeline

Change-Id: I20f84bc20807378ecc7357083dfa2ec4dc05af4f