Hide selected fieldsets in artifact view
This is part of story #12188 hide fieldsets according to state

On a given state, if you manually define a hidden fieldsets post action
in the database, then the given fieldesets will be hidden in any case.

For now, there is no button to show all fieldsets, and fieldsets are
always hidden without taking into account mandatory fields or field
dependencies.

This is only working in the artifact view. No information about this
post action is given through REST.

You have to set a feature flag in the local.inc file to have this post
action working: $sys_should_use_hidden_fieldsets_post_actions = 1;

Change-Id: I257e7f6306eee18ac8a75c50ad9dc189315a8f52