Drop image in ckeditor in follow-up comments
You can now drop an image in the follow-up comment of an artifact, and
image will be saved.

Other text fields of the artifact will be done later (I want to validate
the approach beforehand).

If user drop an image and erase it, then it will not be attached to the
artifact, and will be garbage collected later.

Change-Id: Ia5bf897b655abcfa40978363e89369116cd4599d