Replace process_system_events.php by tuleap process-system-event
This is driven by the need of running Tuleap backend (hence system events)
on centos7 containers without systemd but with supervisord+crond.

As there are no dependencies on /usr/bin/php, it's not deployed and the cron
fails because they start with /usr/bin/php mt_rand...

I could have fixed it by replacing the exec by php72 from remi but even me
find that really too ugly (and needless to say that it will break at some
points when we move to php73).

The rational here is to integrate all lonely .php scripts as tuleap commands
and to have the "random delay" as a primary feature of the command.

This is part of request #13394 tuleap-aio should run on centos7

Change-Id: I00f77379fe0348c97cc2608e7dc2306b7a3e0716