request #13408: Display of a tracker table report can be slow due to a confusion of the MySQL optimizer
The query built to retrieve information for the tracker table report now only pass
the last changeset IDs to retrieve and let MySQL do the join by itself with the tracker_artifact
table. In some cases giving both the artifact IDs and the changeset IDs seems to confuse MySQL.

No functionnal change is expected. The display of the tracker table report should in the
worst case takes as much time as of now or, in the best case, be massively improved.

Change-Id: I85fa5fbbc4b29294ab5f2e6351ada375c38a7560