Prevent potential infinite redirection loop when saving the deactivation of deprecated browser warning
This contribution makes sure the page saving the deactivation of the banner can only
be called with a POST request as all state changing actions should.
It also makes sure that if the CSRf verification fails, the user is redirected to an
existing page.

Fix request #13428: Make sure the page /account/disable_ie7_warning.php does not go in a infinite
redirection loop when the CSRF verification fails

Change-Id: I9126b21724760e243adec3ef6719f3471c47aa1f