Remove inline style for images in ckeditor
When you paste a big images in ckeditor, once the artifact is updated,
the look n feel of the image is distorded. In order to avoid that, we
prevent ckeditor to insert magic style attribute (and let the users
enter their own size if needed).

Part of story #12964: drag'n drop images in tracker text areas

Change-Id: Ie651caa265a7fc26a37322b41ff42d223bf172ee