Reinject tus-uploaded-id on form error at artifact submit
Initiate the submission of an artifact.
Drop an image in a text field.
Remove the required attribute of a required field (e.g title).
Submit.

You get an error asking you to fill the title field. Obey.
Submit.

The artifact is submitted.
The text field contains the image.
The file field contains the image as attachment.

Note: artifact update will be done in a dedicated changeset

Part of story #12964: drag'n drop images in tracker text areas

Change-Id: I235c0e505d55144d01a6dc3ad08739e9b91b50c3