request #13436: PHP class names should be referenced by ::class when possible
This contribution replaces and forces the usage of:
 - self::class instead of __CLASS__
 - parent::class instead of get_parent_class()
 - self::class instead of get_class()
 - static::class instead of get_called_class() and get_class($this)

This brings consistency all over the codebase by forcing the usage of modern
class references. It's also a bit faster for the PHP engine.

To test, you can review the code of the sniff [0], apply the new ruleset and allow
to autofix this rule then launch `make phpcbf`. The diff should be identical with
this contribution (with a few exceptions for fixing spaces issue).

[0] https://github.com/slevomat/coding-standard/blob/5.0.4/SlevomatCodingStandard/Sniffs/Classes/ModernClassNameReferenceSniff.php

Change-Id: Iac33d36a56b6c1699ea23f86d9e17728bdc5de50