request #13449: Prevent tracker workflow trigger rule to update a parent artifact that is not in the target tracker
Like the contributions done for request #13383 and request #13409 this contribution
is only about hardening the existing code to make sure we do not add unwanted changesets
in the artifacts.

The issue only happens in hard to reproduce scenarii when trigger rules
are specified in a specific order and the tracker hierarchy is modified
without removing the trigger rules.

Change-Id: Ib213be2c34256b833ef63f797c860ea7d12c88ee