request #13458: Artifact CSV import does not manage empty or absent list value consistently
With this contribution, it is now possible to select the None value of a list
by leaving an empty string for the field (similar to what's being done with
the CSV export).
This contribution also makes sure that the field are set to a default value
when the CSV import is used to create an artifact but not all the fields are
present in the CSV file. This is the same behavior than the REST API.

Now, when an artifact is created through the CSV import without an explicit
status the artifact will be visible in the backlog column of a kanban instead
of being not visible.

Change-Id: I949a15d4a5bd7d130e1b81ab5918e37ce6ef4d66