Duplicate hidden fieldsets actions at transition creation
This is part of story #12188 hide fieldsets according to state

If your workflow is in simple mode and the state you are creating a new
transition contains a hidden fieldsets post action, then this action is
duplicated for the newly created transition.

To have this working, the feature flag $sys_should_use_hidden_fieldsets_post_actions
must be set to 1 in the local.inc file or in database.

Change-Id: If44c273fed243ec9e16b2a0c67a0100758be6753