On switch from Simple to advanced workflow, delete post-actions
This is part of story #12188 hide fieldsets according to state

If you have manually created a hidden fieldsets post action for a
given state, when you switch your workflow from simple to advanced mode,
then the post action is deleted.

The modale warning message is updated if you have set the feature flag.

Change-Id: Ie88d412e8d34e0b1f213c73d665f4f1c0e22fee7