Extract Text field dedicated directive
Part of story #13453: finish drag'n drop images in tracker text areas

First step to integrating the Image drag'n'drop in the angular modal. To
ease reviews, we start by extracting the text field into a separate
directive. Next we'll extract the follow-ups editor.

How to test:
- build the scrum planning app
- No functional changes expected.

Note to reviewer: There is a bug with HTML format and required text
fields [0]. It was already present before and I can't seem to fix it
with the current integration library used (ng.ckeditor) so it'll have to
wait for the future modal rework in Vue.

[0] request #13467 Artifact-modal allows empty required text fields when
format is HTML

Change-Id: I70ba067819afa700e7e0c0c72dc9cf1f37bac653