request #13472: Forbid late static binding with constants
LSB with constants is confusing and makes harder for implementations
to determine which parts of the contract set by the interface/class
needs to be fulfilled. Abstract methods can be used to achieve the same
thing.

This contribution fixes the only usage of it in the codebase and updates
the coding standards to block new usages.

Change-Id: Ibebdc3f48e4335abfbd4474baacc7bb87fbdae35