XML import of hidden fieldsets postaction
This is part of story #12188 hide fieldsets according to state

The hidden fieldsets post action is now importable through XML import.
Here is an exemple of XML postactions that can be used to import this
postaction:

<postaction_hidden_fieldsets>
    <fieldset_id REF="F1"/>
    <fieldset_id REF="F2"/>
</postaction_hidden_fieldsets>

To have this working, the feature flag $sys_should_use_hidden_fieldsets_post_actions
must be set to 1 in the local.inc file or in database.

Change-Id: I37d8c468c26cb52e6737cd375dafead8f50cd68d