Strengthen typehint in saveNewChangeset parameters
This is a refactoring separated from the rest of the contributions so
that it does not clutter the review. Since a parameter will be
introduced in a next changeset, I prefer to clean the place before.

Having stronger typehint does not harm, and does not produce any
functional changes.

Part of story #13453: finish drag'n drop images in tracker text areas

Change-Id: I4f4c724e6d7a538860c8f5d4f284fec38e952e8c