XML import of dropped attachments
Drag 'n drop (or copy image/paste) an image in a text field (in html
mode).
XML export the artifact/tracker/project[0]
XML import the project
=> In the new project, the imported artifact has an image embedded in
the text field and the url of this image is the one of the imported
attachment.

Follow-up comments will have their own commit.

Part of story #13453: finish drag'n drop images in tracker text areas

[0] https://docs.tuleap.org/administration-guide/projects-management/project-export-import.html

Change-Id: Ibb0914d77ba9663b4945729643d19bbf6b93e625