request #14516: Replace jimdo/prometheus_client_php by tuleap/prometheus-client
This will allow us to re-use the same Redis connection between the instrumentation
and other tasks and to have the full control on how the connection is opened and with
which settings.
The library also contains a few fixes for multiple small things and the internal
refactoring will also help if we want to implement type of metrics that are currently
not supported like the summaries.

The way the Prometheus client stores its information has changed a bit
(loosely typed vs strictly typed) and therefore the previous data must
be cleared otherwise the /metrics page crashes. In order to do so, a
command 'prometheus_metrics:clear' is introduced. However to keep things
as simple for administrators, Tuleap tries to cleanup opportunely the
persistent datastore.

Change-Id: I524d4c7cfe21f4bc7bf53a1c76221ff69fb7a70e