Configure Hide Fieldsets in web UI
This is part of story #12188 hide fieldsets according to state

The hidden fieldsets post action can now be retrived and defined in the
state modale in the new workflow view.

To have this working, the feature flag $sys_should_use_hidden_fieldsets_post_actions
must be set to 1 in the local.inc file or in database.

Change-Id: I16c6b48f99c036bcd3142556576b1f9c840e919a