Update to Psalm 3.4.5
Add back the recording of issues in the CI pipelines, Psalm now has
the capability of generating reports with only the errors. This feature
will also be useful to identify the files reporting issues when analyzed
in isolation.

This contribution also comes with the "fixes" for the issues found by this
new version. Most of them are due to Psalm new capability to infer property
types from simple constructors.

Psalm 3.4.5 now longer stop the analysis of a block when an undefined function
is found. As a result a lot of new deprecated functions has been found. The
low risks functions has been updated, the ones with a non negligeable risks
has been left as is and manually suppressed.

Part of request #13292: Integrate vimeo/psalm to the test pipeline

Change-Id: I526ef4344c73ea5631042a5895dac138a655ca8b