Run PHPUnit unit test suites with PHP 7.4 in the nightly pipeline
Some tests were modifying the global state with mock objects resulting
in crashes when @runInSeparateProcess is used. By default, PHPUnit
tries to preserve the global state when starting a new process but
in PHP 7.4 the mock objects generated by Mockery cannot be serialized
because that means serializing ReflectionClass instances which is now
forbidden.
It is expected from a Tuleap unit test to leave the global state
unmodified after its execution. This was supposedly enforced by
the PHPUnit configuration but a setting was missing to make it work as
expected in PHPUnit 6+. The configuration has been fixed to identified
the broken tests and the tests have been fixed.

Part of request #13487: Run Tuleap test suites with PHP 7.4

Change-Id: Idffe140a16b2c8e4674c9b5fa82304a6df0e3c57