Handle drag n drop of images in Artifact modal
Part of story #13453: finish drag'n drop images in tracker text areas

How to test:
- No functional change expected in the image drag and drop for the
Tracker artifact view. Check translations, errors and check that when
you upload an image and then remove it from the editor, it is filtered
out of the file field.

- Rebuild planning v2
- When you have a file field and you can update / submit to it, you
should be able to paste or drop an image in a Follow-up comment or a
text field.
- When you submit, the file field references that image.
- When you remove the image from the textarea before submitting, the
file field does not reference that image anymore (it is filtered before
submit)
- When there is an error, an error message should be shown below the
textarea.

- When there is no file field or there is one that you can't write to,
there should NOT be a help message below the textarea.
- When you try to paste an image anyway, there should be an error
message below the textarea.

Notes to reviewer:
Translation has been really painful for this patch. Please note that:
- If I had kept the "gettext_provider" in ckeditor files, we would have
imported "po" files with 3 different outcomes: vue-gettext,
angular-gettext and node-gettext, which is very painful to configure.
- make generate-po does not extract gettext from Vue components in the
artifact modal. I tried it, and it drops all strings extracted from
angular.
- "translate" attribute does not work out of the box with NgVue.
- For the two points above, I think it's wiser to spend time reworking
the modal in Vue instead of making them work. Translated strings in Vue
will have to be updated "by hand" without tools to extract them.

Change-Id: I3c92a06dd92fa1b55c9d198ed0c867239db34740