History of File field should be exported
Until now, history of File field was not exported. This means that:

Given I dropped a file in a follow-up comment
And updated again the artifact,

When I exported it,

Then file information being only in the last changeset means that the
previous changesets couldn't be fully imported.

Part of story #13453: finish drag'n drop images in tracker text areas

Change-Id: I76eef1fcb616d3a9de6d411d17d5614b65429c10