request #13516: Empty plugin hooks cache file crashes the instance
The plugin hooks cache file is now overwritten when incorrect data
is retrieved from it (e.g. an empty string after a clear caches and
ownership restoration). This make sure we do not end up in a situation
where we are forced to remove manually the file on the server.

Change-Id: I8b0d397743d9f22a205fdbe3586c539bf82260e9