Use a generic script to start MySQL when running the SimpleTest test suite
This change allows the Docker images to define how MySQL is started.
Under CentOS 7+ /usr/bin/service is replaced by systemd so we need to change
the way the MySQL server is started.

This is needed to run the test suite under PHP 7.4 because the packages are not
(and will not) be available for CentOS6.

This is part of request #13487: Run Tuleap test suites with PHP 7.4

Change-Id: I28f0a61ec6c57c154e14c81697aa716029e456a9