Plugin hooks cache can not be rewritten when it's corrupted but have the expected file permissions
With this contribution the cache is now invalidated before being
regenerated. By default, the cache file is supposed to be read-only
so it cannot be rewritten even if it is incorrect.
This contribution also locks the file before writting something in it,
this should limit the occurence of corruption of the file.

request #13516: Empty plugin hooks cache file crashes the instance

Change-Id: I944ef7c092d5bd6eebfc01e0718595229a50a16d