Exception not well handled in Burndown REST representation
This fixes request #13549 Fatal error in modal V3

Reverting to old logical stuff done in this particular context.
Graph field values and exceptions are used to build null attribute of
TimePeriod object ...

This will be correctly handled (in TimeframeBuilder object)
in a dedicated commit where the build in graph context will be added to
the builder object.

Change-Id: Iae5d8081c4ca9a7c696d3a3d3df2ed0c561562f5