Remove \get_server_url() function
The function is marked as deprecated since 2016.
Since the upgrade to Psalm 3.4.5 it is also responsible for
a non negligeable part of the issues found when a subpart of the codebase
is analyzed (see the attached file in the request).

The change affects some site-content files so a note in the deployment
guide will need to be added.

Part of request #13292: Integrate vimeo/psalm to the test pipeline

Change-Id: I3006924885a9950c4952e2a10c3792bf05b8b4a0