Fix UndefinedClass and UndefinedDocblockClass found by Psalm in individual files
The issues are mostly due to an incorrect case in the class name. PHP itself
is not case sensitive but Composer's autoloader is so if the class has not already
been loaded the analysis fails.

Those issues have been found by analyzing each files of the codebase individually
(resulting report is attached to the request).

Part of request #13292: Integrate vimeo/psalm to the test pipeline

Change-Id: I58a463d79f2ac5f946e3b0d2a345278247f69f36