TimePeriod objects must deal with float duration
This is part of story #12537 configure my burndown/up & sprint with a start date & end date

If a float value is provided as a duration, it must deal with it and
consider the part of the day added as a full day. For example, a
duration set as 2.5, the TimePeriod object must consider that the
duration is 3 days.

Change-Id: I4add34555b68e9880e2d8c36a6dfc51d68c7a318