request #13591 Selected value from field involved in field dependency is always the first in the list
Given that I have two list fields (A and B) linked with a field dependency.
In artifact edition, if I can select for B at least two values (B1 and B2) and
I select and save an artifact with B2, when I go to the artifact view I see B2
in read mode but when I switch in edition the field B, B1 will be selected in the list.

This has been introduced by da872bd821419e63a59f20872d617b8268bbd13b

Change-Id: I691e0845ff32c8d0d44f87eac60401de8627e5df