add description of the release with a link to release plane
story #13296 be able to display release data

Add a description of the release if it exists

In all cases, there is a link to the Overview of the release.

The description of the release is get about the semantic,
so you need to configure a new text field in release tracker, and link with the semantic.

How to test:
	- Install plugin release widget
	- Have a release with a description
	- Have a release without description

Expected results:
	- You have a link at the bottom of the description to go to overview of the release
	- If the description exists, it shows
	- If the description doesn't exist, you see only the link

Change-Id: I1f5c6e2ba6d4b4d97be25f398a10fe8cca06b690